move project into package within src
