+++ src/unmonitorr/arrs/__init__.py
@@ -1,0 +1,9 @@
+from .arrbase import HTTPException
+from .radarr import RadarrClient
+from .sonarr import SonarrClient
+
+__all__ = (
+    "HTTPException",
+    "RadarrClient",
+    "SonarrClient",
+)
