+++ src/unmonitorr/config/__init__.py
@@ -1,0 +1,6 @@
+from .config import Config, LogConfig
+
+__all__ = (
+    "Config",
+    "LogConfig",
+)
