+++ src/unmonitorr/log.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import coloredlogs  # type: ignore
 
-from config import LogConfig
+from unmonitorr.config import LogConfig
 
 # setup logging format
 format_string: str = "%(asctime)s | %(module)s | %(levelname)s | %(message)s"
